+++ src/Libraries/OpenTelemetry/JaegerOptions.cs
@@ -1,0 +1,13 @@
+﻿namespace OpenCodeFoundation.OpenTelemetry
+{
+    public class JaegerOptions
+    {
+        public bool Enabled { get; set; }
+
+        public string? ServiceName { get; set; }
+
+        public string? Host { get; set; }
+
+        public int? Port { get; set; }
+    }
+}
