+++ src/Libraries/OpenTelemetry/OpenTelemetryOptions.cs
@@ -5,13 +5,13 @@ namespace OpenCodeFoundation.OpenTelemetry
     public class OpenTelemetryOptions
         : IConfigurationOptions
     {
-        public bool Enabled { get; set; } = false;
+        public bool Enabled { get; set; }
 
         public bool AlwaysOnSampler { get; set; } = true;
 
-        public bool Istio { get; set; } = false;
+        public bool Istio { get; set; }
 
-        public JaegerOptions Jaeger { get; set; } = new JaegerOptions();
+        public JaegerOptions Jaeger { get; set; } = new ();
 
         public void Validate()
         {
@@ -33,16 +33,5 @@ namespace OpenCodeFoundation.OpenTelemetry
                 throw new ArgumentException("Jaeger Host can not be null if Jaeger is enabled");
             }
         }
-    }
-
-    public class JaegerOptions
-    {
-        public bool Enabled { get; set; } = false;
-
-        public string ServiceName { get; set; }
-
-        public string Host { get; set; }
-
-        public int Port { get; set; }
     }
 }
